requirements.txt
datapizza-ai
datapizza-ai-parsers-docling
qdrant-client
easyocr
pandas
numpy
jupyter
python-dotenv
pypdf
PyMuPDF
Pillow>=10.0.0
openai
